More replacements

+++ bbs_plus/src/threshold/threshold_bbs_plus.rs
@@ -153,7 +153,7 @@ impl<E: Pairing> BBSPlusSignatureShare<E> {
             &phase1.e[sig_index_in_batch],
             &phase1.masked_rs[sig_index_in_batch],
             &phase1.masked_signing_key_shares[sig_index_in_batch],
-            sig_index_in_batch,
+            sig_index_in_batch as u32,
             phase2,
         );
         Ok(Self {
