+++ bbs_plus/src/threshold/utils.rs
@@ -39,18 +39,18 @@ pub fn compute_R_and_u<G: AffineRepr>(
     e: &G::ScalarField,
     masked_r: &G::ScalarField,
     masked_signing_key_share: &G::ScalarField,
-    index_in_output: usize,
+    index_in_output: u32,
     phase2: &Phase2Output<G::ScalarField>,
 ) -> (G, G::ScalarField) {
     let R = base.mul(r).into_affine();
     let mut u = *masked_r * (*e + masked_signing_key_share);
     for (_, (a, b)) in &phase2.z_A {
-        u += a[index_in_output];
-        u += b[index_in_output];
+        u += a[index_in_output as usize];
+        u += b[index_in_output as usize];
     }
     for (_, (a, b)) in &phase2.z_B {
-        u += a[index_in_output];
-        u += b[index_in_output];
+        u += a[index_in_output as usize];
+        u += b[index_in_output as usize];
     }
     (R, u)
 }
